+++ functions.php
@@ -2,10 +2,20 @@
 
 load_theme_textdomain( 'axel', get_template_directory() . '/languages' );
 
-add_theme_support( 'post-thumbnails' );
+if ( ! isset( $content_width ) ) $content_width = 900;
+
+function axel_add_editor_styles() {
+  add_editor_style( 'editor-style.css' );
+}
+add_action( 'admin_init', 'axel_add_editor_styles' );
+
+add_theme_support( 'automatic-feed-links' );
+add_theme_support( "custom-background" );
 add_theme_support( 'custom-header' );
 add_theme_support( 'custom-logo' );
+add_theme_support( 'editor-style' );
 add_theme_support( 'html5' );
+add_theme_support( 'post-thumbnails' );
 add_theme_support( 'title-tag' );
 
 register_nav_menus( array(