+++ header.php
@@ -4,11 +4,16 @@
   <meta charset="<?php bloginfo( 'charset' ); ?>">
   <meta http-equiv="x-ua-compatible" content="ie=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <link rel="stylesheet" href="<?php echo get_stylesheet_directory_uri(); ?>/styles/master.css">
+  <?php if ( is_singular() ) wp_enqueue_script( 'comment-reply' ); ?>
   <?php wp_head(); ?>
 </head>
 <body <?php body_class(); ?>>
 
   <header>
+    <a href="<?php echo esc_url( home_url() ); ?>">
+      <img class="logo" src="" alt="Reddog Systems">
+    </a>
     <nav>
       <?php wp_nav_menu(); ?>
     </nav>