+++ search.php
@@ -6,6 +6,7 @@
       while( have_posts() ) : the_post();
         get_template_part( 'template-parts/post/content' );
       endwhile;
+      the_posts_pagination()
     endif;
   ?>
 </main>