+++ searchform.php
@@ -1,4 +1,4 @@
-<form role="search" method="get" action="<?php bloginfo( 'url' ); ?>">
+<form role="search" method="get" action="<?php echo esc_url( home_url() ); ?>">
 	<input type="search" name="s" />
 	<button type="submit"><?php _e( 'Search', 'axel' ); ?></button>
 </form>